+++ src/Cranky/CrankyFileBuffer.cpp
@@ -3,10 +3,10 @@
 #include "dolphin/os.h"
 #include "macros.h"
 
-void CrankyFileBuffer::SetFileBuffer(u8* buf, int requested_size) {
+void CrankyFileBuffer::SetFileBuffer(u32* buf, int requested_size) {
     int size;
 
-    m_pBuffer = (u8*)CrankyRoundUp32((int)&buf);
+    m_pBuffer = (u32*)CrankyRoundUp32((u32)&buf);
     size = requested_size - (m_pBuffer - buf);
     m_BufferSize = CrankyAlign32(size);
     ASSERTLINE(107, (m_pBuffer + m_BufferSize) <= (buf + size));
